// backend/src/models/farm.model.ts
// Entity type is generated by Prisma — import Farm from '@prisma/client'

export interface CreateFarmDto {
  name: string;
  latitude: number;
  longitude: number;
  size_ha: number;
  soil_type?: string;
}

export interface UpdateFarmDto {
  name?: string;
  latitude?: number;
  longitude?: number;
  size_ha?: number;
  soil_type?: string;
}